requirements.txt
# Core dependencies (required)
PyMuPDF>=1.23.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0

# Gemini AI backend (optional: pip install google-genai python-dotenv)
# google-genai>=1.0.0
# python-dotenv>=1.0.0

# Ollama AI backend (optional: pip install ollama)
# ollama>=0.4.0

# PDF extras (optional)
# pikepdf>=8.0.0        # PDF decryption
# pdf2image>=1.16.0     # PDF vision mode (also needs poppler system package)
